# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Technical Eigen",
        "description": "Dokumentasi API Eigen menggunakan Swagger",
        "version": "1.0"
    },
    "servers":[
        {
            "url": "http://localhost:8000/api",
            "description": "API"
        }
    ],
...
    "paths":{
...
        "/pinjam/{id_member}":{
            "post":{
                "tags":["Pinjam"],
                "summary": "API untuk pinjam buku",
                "parameters":[
                    {
                        "in": "path",
                        "name": "id_member",
                        "schema":{
                            "type":"integer"
                        }
                    }
                ],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "type": "object",
                                "properties":{
                                    "id_book":{
                                        "type":"integer",
                                        "format": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"sukses"
                    }
                }
            }
        },
        "/kembali/{id_member}":{
            "post":{
                "tags":["Pinjam"],
                "summary": "API untuk Pengembalian buku",
                "parameters":[
                    {
                        "in": "path",
                        "name": "id_member",
                        "schema":{
                            "type":"integer"
                        }
                    }
                ],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "type": "object",
                                "properties":{
                                    "id_book":{
                                        "type":"integer",
                                        "format": "integer"
                                    },
                                    "return_date":{
                                        "type":"string",
                                        "format":"date"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"sukses"
                    }
                }
            }
        }
    }
}